+++ python/fast_mlsirm/__init__.py
@@ -22,7 +22,7 @@ from .linking import link_fixed_item_parameters as link_fixed_item_parameters
 from .linking import irt_link as irt_link, IrtLinkResult as IrtLinkResult
 from .equating import equate_observed_scores as equate_observed_scores, equate_neat as equate_neat, EquateResult as EquateResult, equate_observed_scores_kernel as equate_observed_scores_kernel, loglinear_smooth as loglinear_smooth, equate_neat_linear as equate_neat_linear, equating_standard_errors as equating_standard_errors
 from .rt import fit_response_times as fit_response_times, RtFit as RtFit, fit_speed_accuracy as fit_speed_accuracy, rt_person_fit as rt_person_fit
-from .cdm import fit_cdm as fit_cdm, CdmFit as CdmFit, fit_gdina as fit_gdina, GdinaFit as GdinaFit
+from .cdm import fit_cdm as fit_cdm, CdmFit as CdmFit, fit_gdina as fit_gdina, GdinaFit as GdinaFit, validate_q_matrix as validate_q_matrix, QMatrixValidation as QMatrixValidation
 from .mixture import fit_mixture as fit_mixture, MixtureFit as MixtureFit
 from .mixed import fit_mixed_items as fit_mixed_items, MixedFormatFit as MixedFormatFit, MixedItemParameters as MixedItemParameters
 from .lltm import fit_lltm as fit_lltm, LltmFit as LltmFit
@@ -96,6 +96,8 @@ __all__ = [
     "CdmFit",
     "fit_gdina",
     "GdinaFit",
+    "validate_q_matrix",
+    "QMatrixValidation",
     "fit_mixture",
     "MixtureFit",
     "fit_mixed_items",